Implement "limit-last" mode for multipleOptions
We're flipping the order of the arguments array before and after running
through it in order to retroactively inject stuff into result.argv. This
is a SERIOUSLY suboptimal approach, but because I don't predict that the
"limit-last" option will be used that often, a sloppy approach is better
than blowing days of effort over a "clean" approach.